refactor dir to reuse code
